# fe/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "my-fam-tree API",
    "description": "Family platform API",
    "license": {
      "name": "BUSL-1.1",
      "identifier": "BUSL-1.1"
    },
    "version": "0.1.4"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin"
    }
  ],
  "paths": {
...
    "/api/v1/persons": {
      "get": {
        "tags": [
          "persons"
        ],
        "operationId": "persons_list",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "description": "Resume after this person id",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size, 1..=100, default 50",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Family members page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonsListResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "No session"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      },
      "post": {
        "tags": [
          "persons"
        ],
        "operationId": "persons_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonCreateReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Person created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonViewResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "No session"
          },
          "422": {
            "description": "Validation failed"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      }
    },
    "/api/v1/persons/{id}": {
      "get": {
        "tags": [
          "persons"
        ],
        "operationId": "persons_get_one",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Person id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Person",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonViewResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "No session"
          },
          "404": {
            "description": "Not found in this family"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      },
      "delete": {
        "tags": [
          "persons"
        ],
        "operationId": "persons_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Person id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Person deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "No session"
          },
          "403": {
            "description": "Admin or owner required"
          },
          "404": {
            "description": "Not found in this family"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      },
      "patch": {
        "tags": [
          "persons"
        ],
        "operationId": "persons_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Person id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonUpdateReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Person updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonViewResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "No session"
          },
          "403": {
            "description": "Not allowed to edit this person"
          },
          "404": {
            "description": "Not found in this family"
          },
          "422": {
            "description": "Validation failed"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NullResponseBody": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {},
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta"
              }
            ]
          }
        }
      },
      "Pagination": {
        "type": "object",
        "description": "Cursor-based pagination block. `next_cursor` is `None` on the last page.",
        "required": [
          "limit",
          "returned"
        ],
        "properties": {
          "limit": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "returned": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "PersonCreateReq": {
        "type": "object",
        "required": [
          "given_name"
        ],
        "properties": {
          "birth_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "birth_place": {
            "type": "string"
          },
          "death_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "family_name": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "given_name": {
            "type": "string"
          },
          "linked_user_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "name_at_birth": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "PersonUpdateReq": {
        "type": "object",
        "description": "Partial update. Every field is optional; only `Some(_)` fields overwrite\nthe corresponding column. An empty body is rejected as `validation.value_required`.",
        "properties": {
          "birth_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "birth_place": {
            "type": [
              "string",
              "null"
            ]
          },
          "death_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "family_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "gender": {
            "type": [
              "string",
              "null"
            ]
          },
          "given_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "linked_user_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "name_at_birth": {
            "type": [
              "string",
              "null"
            ]
          },
          "nickname": {
            "type": [
              "string",
              "null"
            ]
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
      "PersonView": {
        "type": "object",
        "required": [
          "id",
          "family_id",
          "given_name",
          "family_name",
          "name_at_birth",
          "nickname",
          "gender",
          "birth_place",
          "notes"
        ],
        "properties": {
          "birth_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "birth_place": {
            "type": "string"
          },
          "death_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "family_id": {
            "type": "string",
            "format": "uuid"
          },
          "family_name": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "given_name": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_favourite_for_me": {
            "type": "boolean",
            "description": "Per-user favourite mark for the signed-in caller. Always set on the\nsingle-person GET; the list endpoint leaves it `false` to avoid\nfanning N+1 favourite lookups across the page (the FE re-fetches the\nper-person GET when it opens the drawer)."
          },
          "linked_user_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "name_at_birth": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "photo_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Time-limited presigned URL for the person's photo, or `null` when\nthe person has no photo. Re-presigned on every read; do NOT cache."
          }
        }
      },
      "PersonViewResponseBody": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PersonView"
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta"
              }
            ]
          }
        }
      },
      "PersonsListResponseBody": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PersonView"
            }
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta"
              }
            ]
          }
        }
      },
...
      "ResponseMeta": {
        "type": "object",
        "description": "Per-response metadata. Currently carries pagination cursors, the\n`request_id` for tracing-friendly client diagnostics, and any soft\nvalidation `warnings` attached by the handler.",
        "properties": {
          "pagination": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Pagination"
              }
            ]
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Warning"
            },
            "description": "Soft warnings. Serialized only when non-empty so the wire shape\nstays unchanged for handlers that don't produce any."
          }
        }
      },
...
      "Warning": {
        "type": "object",
        "description": "Soft-validation warning.\n\nSurfaced under `meta.warnings` on the success envelope when the\ncaller's input violates a heuristic rule that we don't want to reject\noutright (e.g. parent-child age gap < 14 years). The `code` is a\nstable i18n key (`warning.<snake_case>`); `path` is an optional JSON\nPointer to the relevant field.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "cookie_access": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access"
      }
    }
  },
...
}